+++ src/transformation/mesh_intersection/triangle_triangle_intersection.rs
@@ -8,9 +8,6 @@ use crate::utils::WBasis;
 #[derive(Copy, Clone, Debug, Default)]
 pub struct TriangleTriangleIntersectionPoint {
     pub p1: Point<Real>,
-    pub p2: Point<Real>,
-    pub f1: FeatureId,
-    pub f2: FeatureId,
 }
 
 #[derive(Clone, Debug)]
@@ -35,8 +32,8 @@ pub fn triangle_triangle_intersection(
     tri1: &Triangle,
     tri2: &Triangle,
 ) -> Option<TriangleTriangleIntersection> {
-    let normal1 = tri1.normal()?;
-    let normal2 = tri2.normal()?;
+    let normal1 = tri1.robust_normal();
+    let normal2 = tri2.robust_normal();
 
     if let Some(intersection_dir) = normal1.cross(&normal2).try_normalize(1.0e-6) {
         let mut range1 = [
@@ -99,82 +96,16 @@ pub fn triangle_triangle_intersection(
             return None;
         }
 
-        let edge_between = |a, b| match (a, b) {
-            (0, 1) | (1, 0) => FeatureId::Edge(0),
-            (1, 2) | (2, 1) => FeatureId::Edge(1),
-            (2, 0) | (0, 2) => FeatureId::Edge(2),
-            _ => FeatureId::Edge(a),
-        };
-
-        let inter_f1 = match (range1[0].2, range1[1].2) {
-            (FeatureId::Vertex(a), FeatureId::Vertex(b)) => edge_between(a, b),
-            (FeatureId::Vertex(v), FeatureId::Edge(e))
-            | (FeatureId::Edge(e), FeatureId::Vertex(v)) => {
-                if e == (v + 1) % 3 {
-                    FeatureId::Face(0)
-                } else {
-                    FeatureId::Edge(e)
-                }
-            }
-            _ => FeatureId::Face(0),
-        };
-        let inter_f2 = match (range2[0].2, range2[1].2) {
-            (FeatureId::Vertex(a), FeatureId::Vertex(b)) => edge_between(a, b),
-            (FeatureId::Vertex(v), FeatureId::Edge(e))
-            | (FeatureId::Edge(e), FeatureId::Vertex(v)) => {
-                if e == (v + 1) % 3 {
-                    FeatureId::Face(0)
-                } else {
-                    FeatureId::Edge(e)
-                }
-            }
-            _ => FeatureId::Face(0),
-        };
-
         let a = if range2[0].0 > range1[0].0 + EPS {
-            TriangleTriangleIntersectionPoint {
-                p1: range2[0].1,
-                p2: range2[0].1,
-                f1: inter_f1,
-                f2: range2[0].2,
-            }
-        } else if range2[0].0 < range1[0].0 - EPS {
-            TriangleTriangleIntersectionPoint {
-                p1: range1[0].1,
-                p2: range1[0].1,
-                f1: range1[0].2,
-                f2: inter_f2,
-            }
+            TriangleTriangleIntersectionPoint { p1: range2[0].1 }
         } else {
-            TriangleTriangleIntersectionPoint {
-                p1: range1[0].1,
-                p2: range2[0].1,
-                f1: range1[0].2,
-                f2: range2[0].2,
-            }
+            TriangleTriangleIntersectionPoint { p1: range1[0].1 }
         };
 
         let b = if range2[1].0 < range1[1].0 - EPS {
-            TriangleTriangleIntersectionPoint {
-                p1: range2[1].1,
-                p2: range2[1].1,
-                f1: inter_f1,
-                f2: range2[1].2,
-            }
-        } else if range2[1].0 > range1[1].0 + EPS {
-            TriangleTriangleIntersectionPoint {
-                p1: range1[1].1,
-                p2: range1[1].1,
-                f1: range1[1].2,
-                f2: inter_f2,
-            }
+            TriangleTriangleIntersectionPoint { p1: range2[1].1 }
         } else {
-            TriangleTriangleIntersectionPoint {
-                p1: range1[1].1,
-                p2: range2[1].1,
-                f1: range1[1].2,
-                f2: range2[1].2,
-            }
+            TriangleTriangleIntersectionPoint { p1: range1[1].1 }
         };
 
         Some(TriangleTriangleIntersection::Segment { a, b })
@@ -216,27 +147,17 @@ pub fn triangle_triangle_intersection(
             crate::transformation::convex_polygons_intersection(&poly1, &poly2, |pt1, pt2| {
                 let intersection = match (pt1, pt2) {
                     (Some(loc1), Some(loc2)) => {
-                        let (f1, p1) = convert_loc(loc1, pts1);
-                        let (f2, p2) = convert_loc(loc2, pts2);
-                        TriangleTriangleIntersectionPoint { p1, p2, f1, f2 }
+                        let (_f1, p1) = convert_loc(loc1, pts1);
+                        let (_f2, _p2) = convert_loc(loc2, pts2);
+                        TriangleTriangleIntersectionPoint { p1 }
                     }
                     (Some(loc1), None) => {
-                        let (f1, p1) = convert_loc(loc1, pts1);
-                        TriangleTriangleIntersectionPoint {
-                            p1,
-                            p2: p1,
-                            f1,
-                            f2: FeatureId::Face(0),
-                        }
+                        let (_f1, p1) = convert_loc(loc1, pts1);
+                        TriangleTriangleIntersectionPoint { p1 }
                     }
                     (None, Some(loc2)) => {
-                        let (f2, p2) = convert_loc(loc2, pts2);
-                        TriangleTriangleIntersectionPoint {
-                            p1: p2,
-                            p2,
-                            f1: FeatureId::Face(0),
-                            f2,
-                        }
+                        let (_f2, p2) = convert_loc(loc2, pts2);
+                        TriangleTriangleIntersectionPoint { p1: p2 }
                     }
                     (None, None) => unreachable!(),
                 };