+++ src/utils/mod.rs
@@ -30,7 +30,7 @@ pub(crate) use self::sort::sort2;
 pub(crate) use self::sort::sort3;
 pub use self::sorted_pair::SortedPair;
 #[cfg(all(feature = "dim3", feature = "std"))]
-pub(crate) use self::spade::sanitize_point;
+pub(crate) use self::spade::sanitize_spade_point;
 pub(crate) use self::weighted_value::WeightedValue;
 pub(crate) use self::wops::{simd_swap, WBasis, WCross, WSign};