+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 ### Modified
 
 - `TypedShape::Custom(u32)` is now `TypedShape::Custom(&dyn Shape)`.
+- `Real` is now exposed through a `use` statement,
+  so that an indirection is removed in documentation:
+  previous occurrences of `Real` now show `f32` or `f64`.
 - Significantly improved the general stability of mesh/mesh intersection calculation.
 
 ## v0.16.1